Ensure all ADK test suites ignore thought signature
